% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{kcount}
\alias{kcount}
\title{Count k-mers.}
\usage{
kcount(x, k = 5, residues = NULL, gap = "-")
}
\arguments{
\item{x}{a matrix of aligned sequences, a list of unaligned sequences,
or a vector representing a single sequence.
Accepted modes are "character" and "raw" (the latter is
for "DNAbin" and "AAbin" objects).}

\item{k}{integer representing the k-mer size. Defaults to 5.
Note that high values of k
may be slow to compute and use a lot of memory due to the large numbers
of calculations required, particularly when the residue alphabet is
also large.}

\item{residues}{either NULL (default; the residue alphabet is automatically
detected from the sequences), a case sensitive character vector
specifying the residue alphabet, or one of the character strings
"RNA", "DNA", "AA", "AMINO". Note that the default option can be slow for
large lists of character vectors. Specifying the residue alphabet is therefore
recommended unless x is a "DNAbin" or "AAbin" object.}

\item{gap}{the character used to represent gaps in the alignment matrix
(if applicable). Ignored for \code{"DNAbin"} or \code{"AAbin"} objects.
Defaults to "-" otherwise.}
}
\value{
a matrix of k-mer counts with one row for each sequence and n^k
  columns (where n is the size of the residue alphabet and k is the
  k-mer size)
}
\description{
Count all k-letter words in a sequence or set of sequences by
  sliding a window of length k over the sequence(s).
}
\details{
TBA
}
\examples{
  ## compute a matrix of k-mer counts for the woodmouse data (ape package)
  ## with a k-mer size of 3
  library(ape)
  data(woodmouse)
  x <- kcount(woodmouse, k = 3)
  ## 15 row matrix with 64 columns for nucleotide 3-mers AAA, AAC, ... TTT
  ##
  ## convert to AAbin object and repeat the operation
  y <- kcount(ape::trans(woodmouse, 2), k = 2)
  ## 15 row matrix with 400 columns for amino acid 2-mers AA, AB, ... , YY
}
\references{
TBA
}
\seealso{
\code{\link{kdistance}} for distance matrices based on k-mer counting
}
\author{
Shaun Wilkinson
}
