% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{kdistance}
\alias{kdistance}
\title{K-mer distance matrix computation.}
\usage{
kdistance(x, k = 5, method = "edgar", residues = NULL, gap = "-", ...)
}
\arguments{
\item{x}{a matrix of aligned sequences or a list of unaligned sequences.
Accepted modes are "character" and "raw" (the latter is
for "DNAbin" and "AAbin" objects).}

\item{k}{integer representing the k-mer size to be used for calculating
the distance matrix. Defaults to 5. Note that high values of k
may be slow to compute and use a lot of memory due to the large numbers
of calculations required, particularly when the residue alphabet is
also large.}

\item{method}{a character string giving the k-mer distance measure
to be used. Currently the available options are \code{"edgar"} (default;
see Edgar (2004) for details) and the standard methods avaialable for
the base function "dist" ("euclidean", "maximum", "manhattan", "canberra",
"binary" and "minkowski").}

\item{residues}{either NULL (default; the residue alphabet is automatically
detected from the sequences), a case sensitive character vector
specifying the residue alphabet, or one of the character strings
"RNA", "DNA", "AA", "AMINO". Note that the default option can be slow for
large lists of character vectors. Specifying the residue alphabet is therefore
recommended unless x is a "DNAbin" or "AAbin" object.}

\item{gap}{the character used to represent gaps in the alignment matrix
(if applicable). Ignored for \code{"DNAbin"} or \code{"AAbin"} objects.
Defaults to "-" otherwise.}

\item{...}{further arguments to be passed to \code{"as.dist"}.}
}
\value{
a distance matrix of class \code{"dist"}
}
\description{
Computes the matrix of k-mer distances between all pairwise comparisons
  of a set of sequences.
}
\details{
this function computes the N x N k-mer distance matrix,
  where N is the number of sequences in the dataset.
}
\examples{
  ## compute the k-mer distance for the woodmouse dataset in the ape package
  ## with a k-mer size of 5
  library(ape)
  data(woodmouse)
  ## trim gappy ends for global alignment
  woodmouse <- woodmouse[, apply(woodmouse, 2, function(v) !any(v == 0xf0))]
  woodmouse.dist <- kdistance(woodmouse, k = 5)
  ## cluster and plot UPGMA tree
  woodmouse.tree <- as.dendrogram(hclust(woodmouse.dist, "average"))
  plot(woodmouse.tree)
}
\references{
Edgar RC (2004) Local homology recognition and distance measures in
  linear time using compressed amino acid alphabets.
  \emph{Nucleic Acids Research}, \strong{32}, 380-385.
}
\seealso{
\code{\link{mbed}} for leaner distance matrices
}
\author{
Shaun Wilkinson
}
