% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mbed}
\alias{mbed}
\title{Convert sequences to vectors of distances to \emph{n} seed sequences.}
\usage{
mbed(x, seeds = NULL, k = 5, residues = NULL, gap = "-",
  counts = FALSE)
}
\arguments{
\item{x}{a matrix of aligned sequences or a list of unaligned sequences.
Accepted modes are "character" and "raw" (the latter is for "DNAbin"
and "AAbin" objects).}

\item{seeds}{optional integer vector indicating which sequences should
be used as the seed sequences. If \code{seeds = NULL} a set of
log(N, 2)^2 non-identical sequences is randomly selected from the
sequence set (where N is the number of sequences; see Blacksheilds et al.
2010). Alternatively, if \code{seeds = 'all'} a standard N x N
distance matrix is computed.}

\item{k}{integer representing the k-mer size to be used for calculating
the distance matrix. Defaults to 5. Note that high values of k
may be slow to compute and use a lot of memory due to the large numbers
of calculations required, particularly when the residue alphabet is
also large.}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the sequences), a case sensitive character vector
specifying the residue alphabet, or one of the character strings
"RNA", "DNA", "AA", "AMINO". Note that the default option can be slow for
large lists of character vectors. Specifying the residue alphabet is therefore
recommended unless x is a "DNAbin" or "AAbin" object.}

\item{gap}{the character used to represent gaps in the alignment matrix
(if applicable). Ignored for \code{"DNAbin"} or \code{"AAbin"} objects.
Defaults to "-" otherwise.}

\item{counts}{logical indicating whether the (usually large) matrix of
k-mer counts should be returned as an attribute of the returned
object. Defaults to FALSE to avoid excessive memory usage.}
}
\value{
returns a N x log(N, 2)^2 matrix of class "mbed" (where N is the
  number of sequences). The returned
  object has additional attributes including a the 'seeds' vector and
  a matrix of k-mer counts with one row for each sequence and 4^k columns.
}
\description{
The \code{mbed} function takes a list of sequences and returns a matrix of
  distances to a subset of seed sequences using the method outlined
  in Blacksheilds et al. (2010).
}
\details{
This function deals with ambiguities by assigning counts proportionally.
  For example the motif ACRTG would assign the 5-mers ACATG and ACGTG counts of
  0.5 each. This algorithm is O N * 4^k in memory and time complexity so can be very
  slow and memory hungry for larger values of k. Further details of the embedding
  process can be found in Blacksheilds et al. (2010). The k-mer distance measure
  used is that of Edgar (2004).
}
\examples{
  library(ape)
  data(woodmouse)
  ## randomly select three sequences as seeds
  set.seed(999)
  seeds <- sample(1:15, size = 3)
  ## embed the woodmouse dataset in three dimensions
  woodmouse.mbed <- mbed(woodmouse, seeds = seeds, k = 5)
  ## print the distance matrix without attributes
  print(woodmouse.mbed[,], digits = 2)
}
\references{
Blackshields G, Sievers F, Shi W, Wilm A, Higgins DG (2010) Sequence embedding
  for fast construction of guide trees for multiple sequence alignment.
  \emph{Algorithms for Molecular Biology}, \strong{5}, 21.
  Edgar RC (2004) Local homology recognition and distance measures in
  linear time using compressed amino acid alphabets.
  \emph{Nucleic Acids Research}, \strong{32}, 380-385.
}
\seealso{
\code{\link{kdistance}} for full N x N distance matrix computation
}
\author{
Shaun Wilkinson
}
