% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mbed}
\alias{print.mbed}
\title{Print summary methods.}
\usage{
\method{print}{mbed}(x, ...)
}
\arguments{
\item{x}{object of various classes.}

\item{...}{additional arguments to be passed between methods.}
}
\value{
NULL (invisibly)
}
\description{
Print summary methods.
}
\author{
Shaun Wilkinson
}
