% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Remove tree nodes with regular expression pattern matching.}
\usage{
prune(tree, pattern, invert = FALSE, untag = FALSE)
}
\arguments{
\item{tree}{an object of class \code{"dendrogram"}.}

\item{pattern}{a regular expression provided as a single string variable.}

\item{invert}{logical indicating whether the branches whose labels match
the regular expression provided in "pattern" should be
discarded (FALSE; default) or kept.
Nodes with no "label" attributes are ignored.}

\item{untag}{logical (used only when invert = TRUE). Indicates whether
the specified pattern should be removed from the branch labels in
the returned object.}
}
\value{
an object of class \code{"dendrogram"}.
}
\description{
\code{"prune"} takes an object of class \code{"dendrogram"} and
  removes all branches whose branch labels match a given regular
  expression.
}
\details{
TBA.
}
\examples{
  newick <- "(A:0.1,B:0.2,(C:0.3,D:0.4):0.5);"
  dendro <- read.dendrogram(text = newick)
  plot(dendro, horiz = TRUE)
  dendro <- prune(dendro, pattern = "^A$")
  plot(dendro, horiz = TRUE)

}
\seealso{
The \code{\link[ape]{drop.tip}} function in the
  \code{\link[ape]{ape}} package performs a similar operation for objects
  of class \code{"phylo"}.
}
\author{
Shaun Wilkinson
}
