% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdown.R
\name{topdown}
\alias{topdown}
\title{Top down phylogenetic trees.}
\usage{
topdown(x, k = 5, residues = NULL, gap = "-", ...)
}
\arguments{
\item{x}{a list or matrix of sequences, possibly an object of class
\code{"DNAbin"} or \code{"AAbin"}. Alternatively can be an object of class
\code{kcounts}.}

\item{k}{integer representing the the k-mer size to be used for calculating
the distance matrix.}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the sequences), a case sensitive character vector
specifying the residue alphabet, or one of the character strings
"RNA", "DNA", "AA", "AMINO". Note that the default option can be slow for
large lists of character vectors. Specifying the residue alphabet is therefore
recommended unless the sequence list is a "DNAbin" or "AAbin" object.}

\item{gap}{the character used to represent gaps in the alignment matrix
(if applicable). Ignored for \code{"DNAbin"} or \code{"AAbin"} objects.
Defaults to "-" otherwise.}

\item{...}{further arguments to be passed to \code{kmeans} (not including
\code{centers}).}
}
\value{
an object of class \code{"dendrogram"}.
}
\description{
Create phylogenetic trees by successively splitting the sequence dataset
  into smaller and smaller subsets.
}
\details{
This function creates phylogenetic trees by successively splitting
  the sequence dataset into smaller and smaller subsets (a.k.a. recursive
  partitioning). This is a "top down" approach to tree-building, as
  opposed to agglomerative "bottom up" methods such as neighbour joining
  and UPGMA. It is useful for large sequence sets (> 20, 000) since the
  usual N x N distance matrix is not required (where N is the number of
  sequences). Instead, an N x log(N, 2)^2
  distance matrix is first derived from the input sequences using the
  \code{\link{mbed}} function of Blacksheilds et al. 2010 and the k-mer
  distance measure of (Edgar 2004). This is done by first selecting
  log(N, 2)^2 'seed' sequences (selected randomly by default)
  and calculating the distance of each sequence to each of the seeds.
  This requires substantially less memory and time than N x N matrix
  computation. The 'embedded' sequences
  are then split recursively using the k-means (clusters = 2) technique.
}
\examples{
\dontrun{
## Cluster the woodmouse dataset from the ape package
library(ape)
data(woodmouse)
## trim gappy ends for global alignment
woodmouse <- woodmouse[, apply(woodmouse, 2, function(v) !any(v == 0xf0))]
## build tree divisively
set.seed(999)
woodmouse.tree <- topdown(woodmouse)
## plot tree
op <- par(no.readonly = TRUE)
par(mar = c(5, 2, 4, 8) + 0.1)
plot(woodmouse.tree, main = "Woodmouse phylogeny", horiz = TRUE)
par(op)
}
}
\references{
Blackshields G, Sievers F, Shi W, Wilm A, Higgins DG (2010) Sequence embedding
  for fast construction of guide trees for multiple sequence alignment.
  \emph{Algorithms for Molecular Biology}, \strong{5}, 21.

  Edgar RC (2004) Local homology recognition and distance measures in
  linear time using compressed amino acid alphabets.
  \emph{Nucleic Acids Research}, \strong{32}, 380-385.
}
\author{
Shaun Wilkinson
}
