% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asphylo.R
\name{as.phylo.dendrogram}
\alias{as.phylo.dendrogram}
\title{Convert dendrogram to "phylo" object.}
\usage{
\method{as.phylo}{dendrogram}(x, ...)
}
\arguments{
\item{x}{a dendrogram object.}

\item{...}{further arguments to be passed between methods.}
}
\value{
an object of class "phylo".
}
\description{
Functions for converting dendrograms to "phylo" objects
  and \emph{vice versa}.
}
\details{
These functions currently work by temporarily writing a tree to Newick
  text and then parsing the string using either \code{\link{read.dendrogram}}
  or \code{\link[ape]{read.tree}}. A faster implementation that avoids the
  transformation to text strings will be available in a future version.
}
\examples{
  newick <- "(A:0.1,B:0.2,(C:0.3,D:0.4):0.5);"
  x <- read.dendrogram(text = newick)
  y <- as.phylo(x)
  z <- as.dendrogram(y)
  identical(x, z)
}
\author{
Shaun Wilkinson
}
