% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder.R
\name{ladder}
\alias{ladder}
\title{Reorder the branches of a dendrogram in a ladderized fashion.}
\usage{
ladder(x, decreasing = FALSE)
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}.}

\item{decreasing}{logical indicating whether the tree should be ladderized upwards
(FALSE; default) or downwards (TRUE).}
}
\description{
\code{"ladder"} takes an object of class "dendrogram" and ladderizes
the branches for easier visual interpretation.
}

