% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Remove tree branches based on regular expression matching.}
\usage{
prune(tree, pattern, keep = FALSE, untag = FALSE)
}
\arguments{
\item{tree}{an object of class \code{"dendrogram"}.}

\item{pattern}{a regular expression provided as a single string variable.}

\item{keep}{logical indicating whether the branches whose labels match the regexp pattern
provided should be kept (TRUE) or discarded (FALSE; default)}

\item{untag}{logical used only when keep = TRUE, indicating whether the pattern should be removed
from the branch labels.}
}
\description{
\code{"prune"} takes an object of class \code{"dendrogram"} and removes all branches whose
branch labels match a given regular expression.
}

