% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ultrametricize}
\alias{ultrametricize}
\title{Make dendrogram ultrametric.}
\usage{
ultrametricize(x)
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}.}
}
\value{
returns an object of class \code{"dendrogram"}.
}
\description{
This is a simple function that sets the 'height' attributes of
  all leaf nodes to  zero.
}
\details{
TBA.
}
\examples{
  newick <- "(A:0.1,B:0.2,(C:0.3,D:0.4):0.5);"
  dendro <- read.dendrogram(text = newick)
  plot(dendro, horiz = T)
  dendro <- ultrametricize(dendro)
  plot(dendro, horiz = T)

}
\author{
Shaun Wilkinson
}
