% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ultrametricize}
\alias{ultrametricize}
\title{Make dendrogram ultrametric.}
\usage{
ultrametricize(x)
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}.}
}
\value{
Returns an object of class \code{"dendrogram"}.
}
\description{
This is a simple function that sets the 'height' attributes of
  all leaf nodes to zero to aid vizualization.
}
\examples{
  x <- read.dendrogram(text = "(A:0.1,B:0.2,(C:0.3,D:0.4):0.5);")
  plot(x, horiz = TRUE)
  x <- ultrametricize(x)
  plot(x, horiz = TRUE)
}
\author{
Shaun Wilkinson
}
